Remove IDLE mode, define new constant
IDLE mode not necessary as it is already a submode of MONITOR

The constants are LONG_INTERVAL and SHORT_INTERVAL. They are used to set
the sensor measurement interval. The short one is used during watering
to see when it is not needed anymore and the long one during the rest of
the time. The aim is to spare the sensor as it degrades quickly with
usage.