Add second simple http server pretending to be an arduino
This is for test purposes in case a working arduino is not at hand.
It obeys the same protocol as will the arduino:
 - PUT: receive data from outside. This is in replacement of anything
received on the Serial interface in case of using the NetworkSerial (not
yet implemented)
 - GET: Doing nothing so far... Could be used to force the arduino to
publish something?

The others are not required
 - POST: Triggers server to send data outside. In the arduino the data
send would be the replacement of any Serial.println() in case of
NetworkSerial