Buffer for payloads are dynamically allocated
They are not needed most of the time. No need to polute memory with so
much unused space.
Allocate instead on the heap when needed.
For efficiency reasons, the payload 0 will stay allocated all the time
on the stack (as this is the one that will be used most of the time in
case of low connection rate). This will avoid memory fragmentation by
constantly creating and deleting it.
In case of higher connection rate the others will be allocated from time
to time when needed. This should lead to minimal fragmentation